refactor(@sounisi5011/encrypted-archive): move core code into `src/core` directory
