+++ packages/encrypted-archive/src/core/cipher.ts
