+++ packages/encrypted-archive/src/core/compress/gzip.ts
