+++ packages/encrypted-archive/src/core/compress/index.ts
