+++ packages/encrypted-archive/src/core/compress/utils.ts
