+++ packages/encrypted-archive/src/core/decrypt.ts
