+++ packages/encrypted-archive/src/core/encrypt.ts
