+++ packages/encrypted-archive/src/core/header/content-identifier.ts
