+++ packages/encrypted-archive/src/core/header/create.ts
