+++ packages/encrypted-archive/src/core/header/index.ts
