+++ packages/encrypted-archive/src/core/header/parser.ts
