+++ packages/encrypted-archive/src/core/header/protocol-buffers-converter/utils.ts
