+++ packages/encrypted-archive/src/core/header/protocol-buffers/header.proto
