+++ packages/encrypted-archive/src/core/header/protocol-buffers/header_pb.d.ts
