+++ packages/encrypted-archive/src/core/header/utils.ts
