+++ packages/encrypted-archive/src/core/nonce.ts
