+++ packages/encrypted-archive/src/core/stream.ts
