+++ packages/encrypted-archive/src/core/types.ts
