+++ packages/encrypted-archive/src/core/utils/index.ts
