+++ packages/encrypted-archive/src/core/utils/stream.ts
