+++ packages/encrypted-archive/src/core/utils/type.ts
