+++ packages/encrypted-archive/tests/compatibility.ts
@@ -7,6 +7,7 @@ import importFrom from 'import-from';
 import semverSatisfies from 'semver/functions/satisfies';
 
 import { version as latestPackageVersion } from '../package.json';
+import { asyncIterable2Buffer } from '../src/core/utils';
 import {
     CompressOptions,
     CryptoAlgorithmName,
@@ -15,7 +16,6 @@ import {
     EncryptOptions,
     KeyDerivationOptions,
 } from '../src/index.node.js';
-import { asyncIterable2Buffer } from '../src/utils';
 import { buffer2chunkArray, isOneOrMoreArray } from './helpers';
 import { optGen } from './helpers/combinations';
