+++ packages/encrypted-archive/tests/unit/key-derivation-function.ts
@@ -3,8 +3,8 @@ import * as util from 'util';
 
 import escapeStringRegexp from 'escape-string-regexp';
 
-import { getKDF, KeyDerivationOptions } from '../../src/key-derivation-function';
-import type { Argon2Options } from '../../src/key-derivation-function/argon2';
+import { getKDF, KeyDerivationOptions } from '../../src/core/key-derivation-function';
+import type { Argon2Options } from '../../src/core/key-derivation-function/argon2';
 import { addNegativeNumber, createDummySizeBuffer, rangeArray } from '../helpers';
 
 /** @see https://github.com/P-H-C/phc-winner-argon2/blob/16d3df698db2486dde480b09a732bf9bf48599f9/include/argon2.h#L57 */
